refactor sth.
