+++ src/uet/oop/bomberman/graphics/GraphicsMGR.java
@@ -22,45 +22,4 @@ public class GraphicsMGR {
         this.gc = gc;
     }
 
-    public void clrscr(Canvas canvas) {
-        gc.clearRect(0, 0, canvas.getWidth(), canvas.getHeight());
-    }
-
-    public void renderWallAndGrass(GameMap gameMap) {
-        for (int i = 0; i < gameMap.getWallAndGrass().size(); i++) {
-            List<Entity> tempMap = gameMap.getWallAndGrass().get(i);
-            tempMap.forEach(g -> g.render(gc));
-        }
-    }
-
-    public void renderItems(GameMap gameMap) {
-        gameMap.getItems().forEach(g -> g.render(gc));
-    }
-
-    public void renderBomber(GameMap gameMap) {
-        gameMap.getPlayer().render(gc);
-    }
-
-    public void renderEnemies(GameMap gameMap) {
-        gameMap.getEnemies().forEach(g -> g.render(gc));
-    }
-
-    public void renderBricks(GameMap gameMap) {
-        gameMap.getBricks().forEach(g -> g.render(gc));
-    }
-
-    public void renderBombList(GameMap gameMap) {
-        for (Entity element : gameMap.getBombList()) {
-            element.render(gc);
-        }
-    }
-
-    public void render(GameMap gameMap) {
-        renderWallAndGrass(gameMap);
-        renderItems(gameMap);
-        renderBricks(gameMap);
-        renderBombList(gameMap);
-        renderEnemies(gameMap);
-        renderBomber(gameMap);
-    }
 }
